+++ src/ClientCore/Rendering/Configuration/DisplayViewport.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Sovereign.ClientCore.Rendering.Configuration
+{
+
+    /// <summary>
+    /// Configures the display viewport used for rendering.
+    /// </summary>
+    public sealed class DisplayViewport
+    {
+
+        /// <summary>
+        /// Width of the display viewport in pixels.
+        /// </summary>
+        public int Width => 1280;
+
+        /// <summary>
+        /// Height of the display viewport in pixels.
+        /// </summary>
+        public int Height => 720;
+
+    }
+
+}