+++ src/D3D11Renderer/Rendering/Scenes/Game/World/WorldOutputMerger.cs
@@ -1,0 +1,99 @@
+﻿/*
+ * Sovereign Engine
+ * Copyright (c) 2018 opticfluorine
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a 
+ * copy of this software and associated documentation files (the "Software"), 
+ * to deal in the Software without restriction, including without limitation 
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense, 
+ * and/or sell copies of the Software, and to permit persons to whom the 
+ * Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER 
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING 
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+ * DEALINGS IN THE SOFTWARE.
+ */
+
+using SharpDX.Direct3D11;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Sovereign.D3D11Renderer.Rendering.Scenes.Game.World
+{
+
+    /// <summary>
+    /// Manages the output-merger stage for world rendering.
+    /// </summary>
+    public sealed class WorldOutputMerger
+    {
+
+        private readonly D3D11Device device;
+
+        /// <summary>
+        /// Blend state.
+        /// </summary>
+        private BlendState blendState;
+
+        public WorldOutputMerger(D3D11Device device)
+        {
+            this.device = device;
+        }
+
+        /// <summary>
+        /// Initializes the output-merger stage.
+        /// </summary>
+        public void Initialize()
+        {
+            var desc = new BlendStateDescription()
+            {
+                AlphaToCoverageEnable = false,
+                IndependentBlendEnable = false
+            };
+            desc.RenderTarget[0] = new RenderTargetBlendDescription()
+            {
+                IsBlendEnabled = true,
+                SourceBlend = BlendOption.SourceAlpha,
+                DestinationBlend = BlendOption.Zero,
+                BlendOperation = BlendOperation.Add,
+                SourceAlphaBlend = BlendOption.One,
+                DestinationAlphaBlend = BlendOption.Zero,
+                AlphaBlendOperation = BlendOperation.Add,
+                RenderTargetWriteMask = ColorWriteMaskFlags.All
+            };
+
+            blendState = new BlendState(device.Device, desc);
+        }
+
+        /// <summary>
+        /// Binds the output-merger stage to the pipeline.
+        /// </summary>
+        /// <param name="context">Device context.</param>
+        public void Bind(DeviceContext context)
+        {
+            context.OutputMerger.SetRenderTargets(device.BackBufferView);
+            context.OutputMerger.SetBlendState(blendState);
+        }
+
+        /// <summary>
+        /// Unbinds the output-merger stage from the pipeline.
+        /// </summary>
+        /// <param name="context">Device context.</param>
+        public void Unbind(DeviceContext context)
+        {
+            context.OutputMerger.SetBlendState(null);
+            context.OutputMerger.SetRenderTargets((RenderTargetView)null);
+        }
+
+    }
+
+}